# api/openapi.yaml
openapi: 3.0.1
info:
  title: VMClarity APIs
  version: 1.0.0
paths:
...
  /scanEstimations:
    get:
      summary: Get all scan estimations. Each scan estimation contains details about a multi-asset scan estimation.
      operationId: GetScanEstimations
      parameters:
        - $ref: '#/components/parameters/odataFilter'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataCount'
        - $ref: '#/components/parameters/odataTop'
        - $ref: '#/components/parameters/odataSkip'
        - $ref: '#/components/parameters/odataExpand'
        - $ref: '#/components/parameters/odataOrderBy'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanEstimations'
        default:
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create a multi-asset scan estimation
      operationId: PostScanEstimations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanEstimation'
        required: true
      responses:
        201:
          description: A new scan estimation was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanEstimation'
        400:
          description: Invalid scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
  /scanEstimations/{scanEstimationID}:
    get:
      summary: Get the details for a given multi-asset scan estimation.
      operationId: GetScanEstimationsScanEstimationID
      parameters:
        - $ref: '#/components/parameters/scanEstimationID'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataExpand'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanEstimation'
        404:
          description: Scan estimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
    put:
      summary: Update a scan estimation.
      operationId: PutScanEstimationsScanEstimationID
      parameters:
        - $ref: '#/components/parameters/scanEstimationID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanEstimation'
        required: true
      responses:
        200:
          description: Updated scan estimation successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanEstimation'
        400:
          description: Invalid scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Scan estimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    patch:
      summary: Patch a scan estimation.
      operationId: PatchScanEstimationsScanEstimationID
      parameters:
        - $ref: '#/components/parameters/scanEstimationID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanEstimation'
        required: true
      responses:
        200:
          description: Patched scan estimation successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanEstimation'
        400:
          description: Invalid scan estimation supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Scan estimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    delete:
      summary: Delete a scan estimation.
      operationId: DeleteScanEstimationsScanEstimationID
      parameters:
        - $ref: '#/components/parameters/scanEstimationID'
      responses:
        200:
          $ref: '#/components/responses/Success'
        404:
          description: Scan estimation ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in all cases of failures.
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in cases of success that returns nothing.
    Metadata:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
...
    Annotations:
      type: array
      description: Generic map of string keys and string values to attach arbitrary non-identifying metadata to objects.
      items:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
...
    ScanTemplate:
      type: object
      properties:
        scope:
          description: |
            The query used to limit the scope of this scan. It uses
            the ODATA $filter query language to limit the collection of assets
            that this scan will operate over. For example
            `startswith(assetInfo.location, 'eu-west-2')` will limit this scan
            to just assets in the eu-west-2 AWS region.
          type: string
        timeoutSeconds:
          type: integer
          minimum: 0
          description: |
            The maximum time in seconds that a scan started from this config
            should run for before being automatically aborted.
        maxParallelScanners:
          type: 'integer'
          minimum: 1
          maximum: 20
          description: "The maximum number of asset scans that can be scheduled in parallel for this scan"
        assetScanTemplate:
          $ref: '#/components/schemas/AssetScanTemplate'
...
    ScanFamiliesConfig:
      type: object
      description: The configuration of the scanner families within a scan config
      properties:
        sbom:
          $ref: '#/components/schemas/SBOMConfig'
        vulnerabilities:
          $ref: '#/components/schemas/VulnerabilitiesConfig'
        malware:
          $ref: '#/components/schemas/MalwareConfig'
        rootkits:
          $ref: '#/components/schemas/RootkitsConfig'
        secrets:
          $ref: '#/components/schemas/SecretsConfig'
        misconfigurations:
          $ref: '#/components/schemas/MisconfigurationsConfig'
        exploits:
          $ref: '#/components/schemas/ExploitsConfig'
        infoFinder:
          $ref: '#/components/schemas/InfoFinderConfig'
    InfoFinderConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    VulnerabilitiesConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    SBOMConfig:
      type: object
      properties:
        enabled:
          type: boolean
        analyzers:
          type: array
          items:
            type: string
    MalwareConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    RootkitsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    SecretsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    MisconfigurationsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
    ExploitsConfig:
      type: object
      properties:
        enabled:
          type: boolean
        scanners:
          type: array
          items:
            type: string
...
    ScannerInstanceCreationConfig:
      type: object
      description: Configuration of scanner instance
      properties:
        useSpotInstances:
          type: boolean
        retryMaxAttempts:
          type: integer
        maxPrice:
          type: string
      required:
        - useSpotInstances
...
    AssetScanTemplate:
      type: object
      properties:
        scanFamiliesConfig:
          $ref: '#/components/schemas/ScanFamiliesConfig'
        scannerInstanceCreationConfig:
          $ref: '#/components/schemas/ScannerInstanceCreationConfig'
...
    ScanEstimations:
      type: object
      properties:
        count:
          description: Total ScanEstimations count according to the given filters
          type: integer
        items:
          description: List of ScanEstimations according to the given filters
          type: array
          items:
            $ref: '#/components/schemas/ScanEstimation'
    ScanEstimation:
      type: object
      allOf:
        - $ref: '#/components/schemas/Metadata'
        - type: object
          properties:
            id:
              type: string
            assetIDs:
              description: List of asset IDs to be estimated
              type: array
              items:
                type: string
            status:
              $ref: '#/components/schemas/ScanEstimationStatus'
            startTime:
              type: string
              format: date-time
            endTime:
              type: string
              format: date-time
            assetScanEstimations:
              description: AssetScanEstimations which make up this ScanEstimation
              type: array
              items:
                $ref: '#/components/schemas/AssetScanEstimationRelationship'
            scanTemplate:
              $ref: '#/components/schemas/ScanTemplate'
            summary:
              $ref: '#/components/schemas/ScanEstimationSummary'
            revision:
              type: integer
            ttlSecondsAfterFinished:
              description: The duration in seconds this resource should last until it is deleted.
              x-go-name: TTLSecondsAfterFinished
              type: integer
            deleteAfter:
              description: The time this resource should be deleted. This value is calculated by endTime + ttlSecondsAfterFinished.
                This should not be set by the user, but use ttlSecondsAfterFinished instead.
              type: string
              format: date-time

    ## TODO fill out the rest of the fields for the ScanEstimationRelationship.
...
    ScanEstimationStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the state of scan estimation.

            | State      | Description                              |
            | ---------- | ---------------------------------------- |
            | Pending    | Initial state for ScanEstimation         |
            | Discovered | Assets have been discovered              |
            | InProgress | Scan estimation is in progress           |
            | Aborted    | ScanEstimation has aborted               |
            | Failed     | ScanEstimation has failed                |
            | Done       | ScanEstimation has finished successfully |
          type: string
          enum:
            - Pending
            - Discovered
            - InProgress
            - Aborted
            - Failed
            - Done
        reason:
          description: |
            Machine readable reason for state transition.

            | State      | Reason              | Description                                                       |
            | ---------- | ------------------- | ----------------------------------------------------------------- |
            | Pending    | Created             | Initial state for ScanEstimation                                  |
            | Discovered | SuccessfulDiscovery | Assets to estimate successfully discovered                        |
            | InProgress | Running             | Scan estimation is currently in progress                          |
            | Aborted    | Cancellation        | ScanEstimation has been cancelled                                 |
            | Failed     | Aborted             | ScanEstimation has failed due to abort                            |
            | Failed     | Error               | ScanEstimation has failed due to an error                         |
            | Failed     | Timeout             | ScanEstimation has failed due to timeout                          |
            | Done       | NothingToEstimate   | ScanEstimation has finished because there was nothing to estimate |
            | Done       | Success             | ScanEstimation has finished successfully                          |
          type: string
          enum:
            - Created
            - SuccessfulDiscovery
            - EstimationsRunning
            - Cancellation
            - Aborted
            - Error
            - TimeOut
            - NothingToEstimate
            - Success
          x-enum-varnames:
            - ScanEstimationStatusReasonCreated
            - ScanEstimationStatusReasonSuccessfulDiscovery
            - ScanEstimationStatusReasonRunning
            - ScanEstimationStatusReasonCancellation
            - ScanEstimationStatusReasonAborted
            - ScanEstimationStatusReasonError
            - ScanEstimationStatusReasonTimeout
            - ScanEstimationStatusReasonNothingToEstimate
            - ScanEstimationStatusReasonSuccess
        message:
          description: Human readable message.
          type: string
        lastTransitionTime:
          description: Last date time when the status has changed.
          type: string
          format: date-time
...
    AssetScanEstimationRelationship:
      type: object
      properties:
        id:
          type: string
...
    ScanEstimationSummary:
      type: object
      description: A summary of the AssetScanEstimations under this ScanEstimation
      properties:
        jobsLeftToRun:
          type: integer
        jobsCompleted:
          type: integer
        totalScanTime:
          description: Total estimated time the scan will take (seconds)
          type: integer
        totalScanSize:
          description: Total estimated size of the scan (GB)
          type: integer
        totalScanCost:
          description: Total estimated cost of the scan ($)
          type: number
          format: float
  responses:
    Success:
      description: Success message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  parameters:
...
    odataFilter:
      name: "$filter"
      in: query
      schema:
        type: string
    odataSelect:
      name: "$select"
      in: query
      schema:
        type: string
    odataCount:
      name: "$count"
      in: query
      schema:
        type: boolean
    odataTop:
      name: "$top"
      in: query
      schema:
        type: integer
    odataSkip:
      name: "$skip"
      in: query
      schema:
        type: integer
    odataExpand:
      name: "$expand"
      in: query
      schema:
        type: string
    odataOrderBy:
      name: "$orderby"
      in: query
      schema:
        type: string
      x-go-name: "OrderBy"
...
    scanEstimationID:
      name: scanEstimationID
      in: path
      required: true
      schema:
        type: string
...
    ifmatch:
      name: If-Match
      in: header
      schema:
        type: integer
...